Create new Copmonent SongContent
